Fix style issues and build error
